fix: handle destroy with missing security group (#381)
This PR makes the following changes:
* Destroying cluser even if security group is missing (for example in
case of ungraceful stop during a previous destroy)

I tested this PR by hand and against incomplete cluster (missing
security group)

+++ flintrock/ec2.py
@@ -195,7 +195,10 @@ class EC2Cluster(FlintrockCluster):
             region=self.region,
             cluster_name=self.name,
         )
-        cluster_group.delete()
+        # Cluster group might already have been killed if a destroy was ungracefully stopped during
+        # a previous execution.
+        if cluster_group:
+            cluster_group.delete()
 
         (ec2.instances
             .filter(
